+++ src/xfrm.cpp
@@ -14,7 +14,16 @@ std::string xfrm::xml()
     os << "<a:xfrm rot=\"" << (int)(this->rot*60000) << "\">";
   }
   os << "<a:off x=\"" << (int)(this->x * 12700) << "\" y=\"" << (int)(this->y * 12700) << "\"/>";
-  os << "<a:ext cx=\"" << (int)(this->width * 12700) << "\" cy=\"" << (int)(this->height * 12700) << "\"/>";
+
+  // The following lines are here to fix bug of pptx file with negative lengths of rectangles and lines
+  // This is a quick-and-dirty solution, which should be handled in some more sophisticated way
+
+  int width = (int)(this->width * 12700);
+  width = width >= 0 ? width : 10;
+  int height = (int)(this->height * 12700);
+  height = height >= 0 ? height : 10;
+
+  os << "<a:ext cx=\"" << width << "\" cy=\"" << height << "\"/>";
   os << "</a:xfrm>";
   return os.str();
 }